requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
transformers==4.35.0
torch==2.1.0+cpu
pydantic==2.5.0
python-multipart==0.0.6
numpy==1.24.3
PyPDF2==3.0.1
python-docx==1.1.0